requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-admin-interface==0.19.1
django-admin-timeline==1.8
django-colorfield==0.7.1
django-debug-toolbar==3.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jet-reboot==1.3.1
django-nine==0.2.5
django-parler==2.3
django-rosetta==0.9.8
feedparser==6.0.10
idna==3.3
Pillow==9.1.1
polib==1.1.1
python-decouple==3.6
requests==2.28.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
